requirements.txt
strands-agents
strands-agents-tools
fastapi
uvicorn[standard]
